// src/constants/enum.ts
/**
 * 通用 Enums
 */
export enum ResponseCode {
  OK = 200,
  MULTIPLE_CHOICE = 300,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  METHOD_NOT_ALLOWED = 405,
  NOT_ACCEPTABLE = 406,
  INTERVAL_SERVER_ERROR = 500,
  OTHER = 600,
}

export enum BOOLEAN_CODE {
  yes = 1,
  not = 0,
}


export const SUCCESS_CODE = [0];